Move channels into own file
